Create different Java classes for the three of the commands
The method in the Parser.java class is becoming a long method,
especially if there is going to be more commands in the future.

Separate Java Classes created per command.

This ensures that each Command class executes their own unique set of
actions to complete that command within their own class.